refactor: extract TypeName rendering and import collection into separate modules
Move rendering logic (to_doc, to_doc_with_lang, render) into
type_name_render.rs and import collection (collect_imports) into
type_name_import.rs. TypeName retains thin delegation methods so
the public API is unchanged.

Closes #19
